Initial lab2 commit

+++ src/main/java/org/salon/dao/ReviewDAO.java
@@ -3,5 +3,5 @@ package org.salon.dao;
 import org.salon.models.Review;
 
 public interface ReviewDAO extends DAO<Review>  {
-    Review getByAppointmentId(int id);
+    Review getByAppointmentId(long id);
 }
